// libs/chats/src/lib/data/interfaces/chats.interface.ts
import {Profile} from "@tt/interfaces/profile";

export interface Chat {
  id: number;
  userFirst: Profile;
  userSecond: Profile;
  messages: Message[];
  companion?: Profile;
}

export interface Message {
  id: number;
  userFromId: number;
  personalChatId: number;
  text: string;
  createdAt: string;
  isRead: boolean;
  updatedAt: string;
  user?: Profile;
  isMine?: boolean;
}

export interface LastMessageRes {
  id: 475;
  userFrom: Profile;
  message: string | null;
  createdAt: string;
  unreadMessages: number;
}